requirements.txt
streamlit==1.26.0
pandas==2.0.3
plotly_express==0.4.1